create Voter component and refactor ArticlePage and CommentCard to use it
